-Fixed so you cant access 'Tavern only' option from other zones. (4).
Changing some game logic from main -> Game.java
