+++ src/Main.java
@@ -25,19 +25,8 @@ public class Main {
         // combatTest(pc, goblin, sc);
 
         //MESSING AROUND WITH ZONES AND TRAVEL
-        travelTest(pc, sc);
+        Game.gameMenu(pc, sc);
     }
-
-    private static void travelTest(PlayerCharacter pc, Scanner sc) {
-        Zone room = new Zone();
-     room.displayCurrentZone(pc); // Temporary intro to the game, change to display intro etc.
-     Utility.promptEnterKey(sc);
-     while (room.checkGameOver() == false) {
-         Utility.clearConsole();
-         room.travelMenu(pc, room);
-         Utility.promptEnterKey(sc);
-     } 
-     }
 
     private static PlayerCharacter setupUser(Scanner sc) {
         System.out.print("What is your name?: ");
